django-projects/Library/pyproject.toml
[tool.poetry]
name = "django-react-project-library"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-ckeditor = "^6.7.1"
django-cors-headers = "^4.4.0"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
django-countries = "^7.6.1"
asgiref = "^3.8.1"
pyjwt = "^2.9.0"
python-decouple = "^3.8"
pytz = "^2024.2"
sqlparse = "^0.5.1"
django-rest-passwordreset = "^1.5.0"
django-rest-knox = "^5.0.2"
django-allauth = "^65.2.0"
django-model-utils = "^5.0.0"
psycopg2-binary = "^2.9.10"
django-debug-toolbar = "^4.4.6"
whitenoise = "^6.8.2"
django-jquery = "^3.1.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
factory-boy = "^3.3.1"
faker = "^33.0.0"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
